+++ core/load_project.py
@@ -148,6 +148,14 @@ class GwLoadProject(QObject):
         lib_vars.project_epsg = tools_qgis.get_epsg()
         tools_gw.connect_signal(QgsProject.instance().crsChanged, tools_gw.set_epsg,
                                 'load_project', 'project_read_crsChanged_set_epsg')
+        tools_gw.disconnect_signal('load_project', 'project_read_willRemoveChildren_protect_vr')
+        tools_qgis.refresh_value_relation_target_tables()
+        tools_gw.connect_signal(
+            QgsProject.instance().layerTreeRoot().willRemoveChildren,
+            tools_gw.protect_value_relation_layers_on_remove,
+            'load_project',
+            'project_read_willRemoveChildren_protect_vr'
+        )
         global_vars.project_loaded = True
 
         # Set indexing strategy for snapping so that it uses less memory if possible